dalliance: Save widget locally and view through localhost.
This is required in order to bypass Cross-Origin Request limitations.
If we just displayed the widget with `createWidget', dalliance.js would
have no way of loading local data files, as they would be served through
the HTTP protocol, not the file:// protocol.

Hence, we decide to serve everything over a local HTTP server --- data
files as well as the actual widget.